requirements.txt
langchain~=0.0.274
chromedriver-py
selenium~=4.11.2
openai~=0.27.8
google-search-results
tiktoken
wikipedia
chromadb
PyPDF2~=3.0.1
sentence_transformers
gradio~=3.41.2
'httpx[socks]'

PyYAML~=6.0.1
requests~=2.31.0
beautifulsoup4~=4.12.2
pydantic~=1.10.12
codeinterpreterapi~=0.0.12